Rewrite find_stalkerware in python, add graphs for controlled exp.
I'll want to add functionality to find_stalkerware soon and I don't
want to do that in bash. Also, running Aaron's controlled experiment,
where we try to put an unused domain in cache and see what happens.
So far that hasn't worked as intended. Maybe we haven't found a
sufficiently unused domain.